[hotfix] Move testing sender and receiver invokables to TestingAbstractInvokables
Moves the testing Sender and Receiver class from JobExecutionITCase to TestingAbstractInvokables
for easier reusability.
